+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - add NIP-38 (user statuses)
 - add NIP-03 (OpenTimestamps attestations)
 - add NIP-37 (draft events)
+- add NIP-25 external reaction (kind 17)
 
 ## v0.1.0 (2023-06-07)